Mark zeroconf.logger as protected by renaming to zeroconf._logger
- The public API should only access zeroconf and zeroconf.aio
  as internals may be relocated between releases
