+++ zeroconf/_handlers.py
@@ -24,6 +24,7 @@ import itertools
 from typing import List, Optional, TYPE_CHECKING, Union
 
 from ._dns import DNSAddress, DNSIncoming, DNSOutgoing, DNSPointer, DNSQuestion, DNSRecord
+from ._logger import log
 from .const import (
     _CLASS_IN,
     _DNS_OTHER_TTL,
@@ -36,7 +37,6 @@ from .const import (
     _TYPE_SRV,
     _TYPE_TXT,
 )
-from .logger import log
 from .services import (
     RecordUpdateListener,
 )
